// src/core/engine/enums/node-type.enum.ts
export enum NodeType {
  PARTIAL = "Partial",
  HTML = "HTML",
  IDENTIFIER = "Identifier",
  EXPRESSION = "Expression",
  IF_STATEMENT = "IfStatement",
  ELSE_STATEMENT = "ElseStatement",
  EACH_STATEMENT = "EachStatement",
  ROOT = "root",
  BOOLEAN_IDENTIFIER = "BooleanIdentifier",
  INTERPOLATION = "Interpolation"
}